# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AnyLive API
  version: 0.1.0
  description: |
    AnyLive overseas-first live streaming control plane.
    P0 surface: system health + meta. Auth/rooms freeze in v0.1/v0.2.
servers:
  - url: http://localhost:8088
    description: Local API
...
paths:
...
  /api/v1/auth/otp/send:
    post:
      tags: [auth]
      operationId: authOtpSend
      summary: Send email OTP (P1)
      description: Stub contract — implemented in auth milestone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendRequest"
      responses:
        "204":
          description: OTP sent (or silently accepted)
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/auth/otp/verify:
    post:
      tags: [auth]
      operationId: authOtpVerify
      summary: Verify OTP and issue tokens (P1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Tokens + user issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/oauth/exchange:
    post:
      tags: [auth]
      operationId: authOauthExchange
      summary: Exchange OAuth id_token for session (WBS E2.1 scaffold)
      description: |
        Local/dogfood: `id_token = "stub:<email>"` when OAUTH_STUB=1 or local APP_ENV.
        Production forbids stub. Real Google/Apple JWKS verification still needs vendor config.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OauthExchangeRequest"
      responses:
        "200":
          description: Tokens + user issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/token/refresh:
    post:
      tags: [auth]
      operationId: authTokenRefresh
      summary: Rotate refresh token (P1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: authLogout
      summary: Revoke refresh token(s) (P1)
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me/sessions:
    get:
      tags: [auth]
      operationId: listMySessions
      summary: List active refresh sessions (jti only; WBS E2.4)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      tags: [auth]
      operationId: logoutAllSessions
      summary: Revoke all refresh sessions for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revoked count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutAllResponse"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me/sessions/{jti}:
    delete:
      tags: [auth]
      operationId: revokeSession
      summary: Revoke a single refresh session by jti (must belong to caller)
      security:
        - bearerAuth: []
      parameters:
        - name: jti
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session revoked
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/me:
    get:
      tags: [users]
      operationId: getMe
      summary: Current user profile (P1)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
    patch:
      tags: [users]
      operationId: patchMe
      summary: Update display name and age/privacy declarations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMeRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      tags: [compliance]
      operationId: deleteMe
      summary: Soft-delete account and revoke refresh tokens
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
    OtpSendRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    OtpVerifyRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 4
          maxLength: 8
    OauthExchangeRequest:
      type: object
      required: [provider, id_token]
      properties:
        provider:
          type: string
          description: google | apple
        id_token:
          type: string
          description: Provider ID token; local dogfood may use stub:<email>
        invite_code:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token lifetime seconds
    AuthSession:
      type: object
      required: [user, access_token, refresh_token, expires_in]
      properties:
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
    Session:
      type: object
      required: [jti, expires_at]
      properties:
        jti:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
    SessionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Session"
    LogoutAllResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: integer
          format: int64
...
    User:
      type: object
      required: [id, display_name, created_at, age_confirmed, privacy_accepted]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        age_confirmed:
          type: boolean
        privacy_accepted:
          type: boolean
        avatar_url:
          type: string
          format: uri
          description: Public avatar object URL when set
        region:
          type: string
          description: Optional ISO 3166-1 alpha-2 region / country code
    PatchMeRequest:
      type: object
      properties:
        display_name:
          type: string
        age_confirmed:
          type: boolean
        privacy_accepted:
          type: boolean
        region:
          type: string
          description: ISO region code; empty string clears
      description: At least one field is required
...